+++ src/redux/sagas/expenditures/createExpenditure.ts
@@ -19,8 +19,8 @@ import {
   initiateTransaction,
   uploadAnnotation,
   getPayoutsWithSlotIds,
-  getPayoutAmount,
   createActionMetadataInDB,
+  getExpenditureValuesMulticallData,
 } from '../utils/index.ts';
 
 export type CreateExpenditurePayload =
@@ -151,44 +151,12 @@ function* createExpenditure({
       ActionTypes.TRANSACTION_CREATED,
     );
 
-    const multicallData: string[] = [];
-    multicallData.push(
-      colonyClient.interface.encodeFunctionData('setExpenditureRecipients', [
-        expenditureId,
-        payoutsWithSlotIds.map((payout) => payout.slotId),
-        payoutsWithSlotIds.map((payout) => payout.recipientAddress),
-      ]),
+    const multicallData = getExpenditureValuesMulticallData(
+      colonyClient,
+      expenditureId,
+      payoutsWithSlotIds,
+      networkInverseFee,
     );
-
-    multicallData.push(
-      colonyClient.interface.encodeFunctionData('setExpenditureClaimDelays', [
-        expenditureId,
-        payoutsWithSlotIds.map((payout) => payout.slotId),
-        payoutsWithSlotIds.map((payout) => payout.claimDelay),
-      ]),
-    );
-
-    const tokenAddresses = new Set(
-      payoutsWithSlotIds.map((payout) => payout.tokenAddress),
-    );
-
-    tokenAddresses.forEach((tokenAddress) => {
-      const tokenPayouts = payoutsWithSlotIds.filter(
-        (payout) => payout.tokenAddress === tokenAddress,
-      );
-      const tokenAmounts = tokenPayouts.map((payout) =>
-        getPayoutAmount(payout, networkInverseFee),
-      );
-
-      multicallData.push(
-        colonyClient.interface.encodeFunctionData('setExpenditurePayouts', [
-          expenditureId,
-          tokenPayouts.map((payout) => payout.slotId),
-          tokenAddress,
-          tokenAmounts,
-        ]),
-      );
-    });
 
     yield put(transactionAddParams(setExpenditureValues.id, [multicallData]));
     yield initiateTransaction({ id: setExpenditureValues.id });
