+++ src/redux/sagas/expenditures/createStakedExpenditure.ts
@@ -16,11 +16,11 @@ import {
   getColonyManager,
   putError,
   takeFrom,
-  getSetExpenditureValuesFunctionParams,
   saveExpenditureMetadata,
   initiateTransaction,
   uploadAnnotation,
   getPayoutsWithSlotIds,
+  getExpenditureValuesMulticallData,
 } from '../utils/index.ts';
 
 export type CreateStakedExpenditurePayload =
@@ -95,7 +95,7 @@ function* createStakedExpenditure({
 
     yield fork(createTransaction, setExpenditureValues.id, {
       context: ClientType.ColonyClient,
-      methodName: 'setExpenditureValues',
+      methodName: 'multicall',
       identifier: colonyAddress,
       group: {
         key: batchKey,
@@ -192,17 +192,15 @@ function* createStakedExpenditure({
       setExpenditureValues.channel,
       ActionTypes.TRANSACTION_CREATED,
     );
-    yield put(
-      transactionAddParams(
-        setExpenditureValues.id,
-        getSetExpenditureValuesFunctionParams(
-          expenditureId,
-          payoutsWithSlotIds,
-          networkInverseFee,
-          isStaged,
-        ),
-      ),
+
+    const multicallData = getExpenditureValuesMulticallData(
+      colonyClient,
+      expenditureId,
+      payoutsWithSlotIds,
+      networkInverseFee,
     );
+
+    yield put(transactionAddParams(setExpenditureValues.id, [multicallData]));
     yield initiateTransaction({ id: setExpenditureValues.id });
     yield waitForTxResult(setExpenditureValues.channel);
